fix: fix profile saving
